+++ hawk/src/androidTest/java/com/orhanobut/hawk/DataHelperTest.java
@@ -25,57 +25,9 @@ public class DataHelperTest extends InstrumentationTestCase {
     context = null;
   }
 
-  public void testGetDataInfoShouldReturnNull_null() {
-    try {
-      DataHelper.getDataInfo(null);
-      assertTrue(false);
-    } catch (Exception e) {
-      assertTrue(true);
-    }
-  }
-
-  public void testGetDataInfoShouldThrowException_NotContainDelimiter() {
-    try {
-      DataHelper.getDataInfo("324234");
-      assertTrue(false);
-    } catch (Exception e) {
-      assertTrue(true);
-    }
-  }
-
-  public void testGetDataInfoShouldReturnNotNull() {
-    DataInfo info = DataHelper.getDataInfo("String11@asdfjasdf");
-    assertNotNull(info);
-  }
-
-  public void testGetDataInfoShouldBeSerializable() {
-    DataInfo info = DataHelper.getDataInfo("String11@asdfjasdf");
-    assertTrue(info.isSerializable());
-  }
-
-  public void testGetDataInfoShouldNotBeSerializable() {
-    DataInfo info = DataHelper.getDataInfo("String10@asdfjasdf");
-    assertFalse(info.isSerializable());
-  }
-
-  public void testGetDataInfoShouldBeList() {
-    DataInfo info = DataHelper.getDataInfo("String11@asdfjasdf");
-    assertTrue(info.isSerializable());
-  }
-
-  public void testGetDataInfoShouldNotBeList() {
-    DataInfo info = DataHelper.getDataInfo("String00@asdfjasdf");
-    assertFalse(info.isSerializable());
-  }
-
   public void testNewVersionCheck() {
-    DataInfo info = DataHelper.getDataInfo("String#String#00V@asdfjasdf");
+    DataInfo info = DataHelper.getDataInfo("java.lang.String##00V@asdfjasdf");
     assertTrue(info.isNewVersion());
-  }
-
-  public void testOldVersionCheck() {
-    DataInfo info = DataHelper.getDataInfo("String#String#00@asdfjasdf");
-    assertTrue(!info.isNewVersion());
   }
 
 }
