+++ hawk/src/androidTest/java/com/orhanobut/hawk/HawkEncryptionMediumTest.java
@@ -3,12 +3,12 @@ package com.orhanobut.hawk;
 /**
  * @author Orhan Obut
  */
-public class HawkSqliteTest extends HawkTest {
+public class HawkEncryptionMediumTest extends HawkTest {
 
   @Override
   public void init() {
     Hawk.init(context)
-        .setStorage(HawkBuilder.newSqliteStorage(context))
+        .setEncryptionMethod(HawkBuilder.EncryptionMethod.MEDIUM)
         .build();
   }
 }
