+++ hawk/src/androidTest/java/com/orhanobut/hawk/HawkNoEncryptionTest.java
@@ -3,12 +3,12 @@ package com.orhanobut.hawk;
 /**
  * @author Orhan Obut
  */
-public class HawkSqliteTest extends HawkTest {
+public class HawkNoEncryptionTest extends HawkTest {
 
   @Override
   public void init() {
     Hawk.init(context)
-        .setStorage(HawkBuilder.newSqliteStorage(context))
+        .setEncryptionMethod(HawkBuilder.EncryptionMethod.NO_ENCRYPTION)
         .build();
   }
 }
