+++ hawk/src/androidTest/java/com/orhanobut/hawk/HawkRxTest.java
@@ -22,7 +22,7 @@ public class HawkRxTest extends InstrumentationTestCase {
         "dexmaker.dexcache",
         getInstrumentation().getTargetContext().getCacheDir().getPath());
     context = getInstrumentation().getContext();
-    Hawk.init(context);
+    Hawk.init(context).build();
   }
 
   @Override
