+++ hawk/src/androidTest/java/com/orhanobut/hawk/HawkSharedPrefsTest.java
@@ -6,30 +6,9 @@ package com.orhanobut.hawk;
 public class HawkSharedPrefsTest extends HawkTest {
 
   @Override
-  protected void setUp() throws Exception {
-    super.setUp();
-    System.setProperty(
-        "dexmaker.dexcache",
-        getInstrumentation().getTargetContext().getCacheDir().getPath());
-    context = getInstrumentation().getContext();
+  public void init() {
     Hawk.init(context)
         .setStorage(HawkBuilder.newSharedPrefStorage(context))
         .build();
-  }
-
-  public void testHugeData() {
-    for (int i = 0; i < 100; i++) {
-      Hawk.put("" + i, "" + i);
-    }
-    assertTrue(true);
-  }
-
-  public void testHugeDataWithBulk() {
-    Hawk.Chain chain = Hawk.chain();
-    for (int i = 0; i < 10000; i++) {
-      chain.put("" + i, "" + i);
-    }
-    chain.commit();
-    assertTrue(true);
   }
 }
